Finished up the search page
